Added GUI for teleportation
Added in the loading of waystones form a player's saved stones into a GUI, with preparations for teleportation.

Will use a trigger system such that -2 is cancel, -1 is no slection, 0+ is the id of the waystone to teleport to.

Trigger will only be enabled in GUI. Any action to leave GUI changes it to something other than -1